+++ cfa/catalog/public/workflows/etl/stf/nssp_gold_deprecated.py
@@ -6,7 +6,7 @@ import tqdm
 from cfa.cloudops.blob_helpers import read_blob_stream, walk_blobs_in_container
 from cfa.dataops import datacat
 
-dataset = datacat.public.stf.nssp_gold
+dataset = datacat.public.stf.nssp_gold_v1
 source_blob = SimpleNamespace(**dataset.config["source"]["storage_location"])
 
 
@@ -27,7 +27,8 @@ def copy_missing_files() -> None:
     files_to_copy = [
         i
         for i in nssp_files
-        if i.removesuffix(".parquet") not in cached_versions
+        if i.endswith(".parquet")
+        and i.removesuffix(".parquet") not in cached_versions
     ]
 
     if not files_to_copy: