+++ changelog.md
@@ -7,6 +7,34 @@ and this project adheres to [Calendar Versioning](https://calver.org/).
 The versioning pattern is `YYYY.MM.DD.micro
 
 ---
+# [2026.04.16.0]
+
+## Changed
+- nssp_gold renamed to nssp_gold_v1
+- new load location in config for nssp_gold_v1
+
+## Added
+- nssp_gold_v2 dataset running in dagster
+
+# [2026.04.09.0]
+
+## Added
+
+- add comprehensive nssp gold (name change) and backfill vintage data function
+- change location of Blob storage
+
+# [2026.03.05.0]
+
+## Added
+
+- add nssp_latest_comprehensive dataset
+
+# [2026.02.23.0]
+
+## Fixed
+
+- fixed nssp pipeline to exclude non-parquet files
+
 # [2026.02.17.0]
 
 ## Added